removes extra line height in popup content
