+++ frontend/src/Components/Event/constants.jsx
@@ -6,6 +6,26 @@ import relativeTime from 'dayjs/plugin/relativeTime';
 
 dayjs.extend(relativeTime);
 
+export const EXISTING_PROJECTS_MSG =
+  'By deactivating the project, all associated content and features will be disabled. This action cannot be undone.';
+
+export const EXISTING_PROJECTS_NOT_FOUND_MSG = 'Hmm.. Seems like you have no active events created.';
+
+export const EVENT_FORM = {
+  eventID: '',
+  reason: '',
+};
+
+export const EVENT_FORM_ERROR = {
+  eventID: '',
+  reason: '',
+};
+
+export const EVENT_FORM_TOUCHED = {
+  eventID: false,
+  reason: false,
+};
+
 export const UNKNOWN_LABEL = 'Unknown';
 export const LABELS = [
   {
